models/reviews-parsing-ner-aspects/requirements.no-gpu.txt
# Copy requirements.txt from MLflow Model Registry
# but remove the GPU-related parts and Torch part to install separately
mlflow==2.14.2
cffi==1.16.0
defusedxml==0.7.1
dill==0.3.8
numpy==1.26.4
packaging==24.1
protobuf==4.25.3
pyyaml==6.0.1
requests==2.32.3
sentencepiece==0.2.0
transformers==4.41.0
# Manually add other packages
mlserver==1.6.0
loguru==0.7.2
# OpenTelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
